Use `.npmignore` to exclude files from release
Exclude files from this repo that are used as part of the build process
and shouldn't be included in a release to npm.

I've tested this locally with `npm pack` and it exludes the files
correctly.

Looking at previously released packages from npm `VERSION.txt` was being
included by the Travis config/dir weren't. I'm not sure why, as npm
publish will include dotfiles if they exist (again, tested locally with
`npm pack`). As we're using the Travis deploy mechanism to actually
publish to npm I suspect they may be manually excluded their own files
as a courtesy. Regardless, we should exclude them here as we have had to
run `npm publish` locally before, and we may move away from Travis for
deployment in the future.