Simplify npm verioning and git committing on publish
The old behaviour did the first 'Temporary' commit to ensure the
the working directoy has no uncommited changes, as `npm version x`
will create a commit after updating the `package.json` version field,
but will fail if there are uncommited local changes.

This requires the `git reset` to go back 2 commits - before the `npm
version` one and the temporary one - but uses `--soft` to not discard
the staged changes from those commits, so they can be commited in the
final `git commit`. Without the reset this script wouldn't push a
single commit, which is more confusing when trying to follow git history.

It works, but it's not easy to follow, especially at first glance.

I've simplified the behaviour by using the `--no-git-tag-version` flag,
which will _only_ update the `package.json` version field, and not try
to create a commit (or a git tag for the version). In the old version
tags weren't being pushed to the remote, so we're not losing anything.

Avoiding the `npm version` commit avoids the need for a temp commit, or
a soft reset, so we can commit once and push that commit to the remote
all in one go, in a much more readable fashion.